[Shared Mount] Add Kernel Param Parsing
