+++ pkg/csi_driver/node.go
@@ -1126,12 +1126,15 @@ func (s *nodeServer) executeNodeStageVolume(ctx context.Context, req *csi.NodeSt
 	}
 
 	// Prepare mount options specifically for shared node mount architecture before sending to mounter pod.
-	preparedMountOptions := prepareSharedNodeMountOptions(args.fuseMountOptions)
+	gcsfuseMO, _, _, err := util.PrepareSharedMountOptions(args.fuseMountOptions)
+	if err != nil {
+		return nil, status.Errorf(codes.InvalidArgument, "failed to prepare mount options: %v", err)
+	}
 
 	// TODO(FUECHR): Pass read_ahead_kb and node_fuse_max_request_limit_kb to gcsfuse config correctly for shared node mounts.
 
 	// Send GRPC to mounter pod to start GCSFuse.
-	if err := s.mountToNode(ctx, podUID, stagingPath, volumeID, preparedMountOptions); err != nil {
+	if err := s.mountToNode(ctx, podUID, stagingPath, volumeID, gcsfuseMO); err != nil {
 		klog.Errorf("Failed to mount volume %q to staging path %q: %v", volumeID, stagingPath, err)
 		if code, fileErr := checkMounterPodErrorFile(emptyDirBasePath); fileErr != nil {
 			return nil, status.Error(code, fileErr.Error())
