+++ pkg/csi_mounter/csi_mounter.go
@@ -26,9 +26,7 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
-	"regexp"
 	"slices"
-	"strconv"
 	"strings"
 	"sync"
 	"syscall"
@@ -37,29 +35,12 @@ import (
 	sidecarmounter "github.com/googlecloudplatform/gcs-fuse-csi-driver/pkg/sidecar_mounter"
 	"github.com/googlecloudplatform/gcs-fuse-csi-driver/pkg/util"
 	"github.com/googlecloudplatform/gcs-fuse-csi-driver/pkg/webhook"
-	"k8s.io/apimachinery/pkg/util/sets"
 	"k8s.io/klog/v2"
 	"k8s.io/mount-utils"
 )
 
 const (
-	// Note: All variables here are in KiB instead of KB but they are added like this to ensure consistency in codebase.
-	socketName                       = "socket"
-	readAheadKBMountFlagRegexPattern = "^read_ahead_kb=(.+)$"
-	readAheadKBMountFlag             = "read_ahead_kb"
-	// Hidden CSI flag never to be passed to gcsfuse
-	nodeFuseMaxRequestLimitKBMountFlag             = "node_fuse_max_request_limit_kb"
-	nodeFuseMaxRequestLimitKBMountFlagRegexPattern = "^" + nodeFuseMaxRequestLimitKBMountFlag + "=(.+)$"
-	defaultNodeFuseMaxRequestLimitKB               = 16 * util.MiB / util.KiB // 16 MiB Request Size
-	// minPageSizeKB defines the minimum page size (4 KiB) to use as a fallback.
-	minPageSizeKB = 4
-	// maxFuseMaxPagesLimit defines the maximum pages limit supported by the Linux FUSE kernel (2^16 - 1).
-	maxFuseMaxPagesLimit = 65535
-)
-
-var (
-	readAheadKBMountFlagRegex               = regexp.MustCompile(readAheadKBMountFlagRegexPattern)
-	nodeFuseMaxRequestLimitKBMountFlagRegex = regexp.MustCompile(nodeFuseMaxRequestLimitKBMountFlagRegexPattern)
+	socketName = "socket"
 )
 
 // Mounter provides the Cloud Storage FUSE CSI implementation of mount.Interface
@@ -90,7 +71,7 @@ func (m *Mounter) Mount(source string, target string, fstype string, options []s
 	m.mux.Lock()
 	defer m.mux.Unlock()
 
-	csiMountOptions, sidecarMountOptions, sysfsBDI, fuseMaxPagesLimit, err := prepareMountOptions(options)
+	csiMountOptions, sidecarMountOptions, sysfsBDI, fuseMaxPagesLimit, err := util.PrepareSidecarMountOptions(options)
 	if err != nil {
 		return err
 	}
@@ -136,7 +117,7 @@ func (m *Mounter) Mount(source string, target string, fstype string, options []s
 	if len(sysfsBDI) != 0 {
 		go func() {
 			if slices.Contains(sidecarMountOptions, fmt.Sprintf("%s=true", util.EnableGCSFuseKernelParams)) {
-				klog.Warningf("The mount option `%s` is soft deprecated for zonal buckets. GCSFuse automatically applies appropriate Read Ahead Kb kernel setting for best performance by default for Zonal Buckets.", readAheadKBMountFlag)
+				klog.Warning("The mount option read_ahead_kb is soft deprecated for zonal buckets. GCSFuse automatically applies appropriate Read Ahead Kb kernel setting for best performance by default for Zonal Buckets.")
 			}
 			// updateSysfsConfig may hang until the file descriptor (fd) is either consumed or canceled.
 			// It will succeed once dfuse finishes the mount process, or it will fail if dfuse fails
@@ -382,93 +363,6 @@ func startAcceptConn(l net.Listener, logPrefix string, msg []byte, fd int, cance
 	}
 
 	klog.V(4).Infof("%v exiting the listener goroutine.", logPrefix)
-}
-
-func prepareMountOptions(options []string) ([]string, []string, map[string]int64, int64, error) {
-	allowedOptions := map[string]bool{
-		"exec":    true,
-		"noexec":  true,
-		"atime":   true,
-		"noatime": true,
-		"sync":    true,
-		"async":   true,
-		"dirsync": true,
-	}
-
-	csiMountOptions := []string{
-		"nodev",
-		"nosuid",
-		"allow_other",
-		"default_permissions",
-		"rootmode=40000",
-		fmt.Sprintf("user_id=%d", os.Getuid()),
-		fmt.Sprintf("group_id=%d", os.Getgid()),
-	}
-
-	// users may pass options that should be used by Linux mount(8),
-	// filter out these options and not pass to the sidecar mounter.
-	validMountOptions := []string{"rw", "ro"}
-	optionSet := sets.NewString(options...)
-	for _, o := range validMountOptions {
-		if optionSet.Has(o) {
-			csiMountOptions = append(csiMountOptions, o)
-			optionSet.Delete(o)
-		}
-	}
-
-	sysfsBDI := make(map[string]int64)
-	pageSizeKB := max(minPageSizeKB, int64(os.Getpagesize()/util.KiB))
-	// Calculate default FUSE max pages limit. We use integer ceiling division to ensure we round
-	// up to the next page if the default size 16MB is not a perfect multiple of the page size on the machine.
-	fuseMaxPagesLimit := util.CeilDiv64(defaultNodeFuseMaxRequestLimitKB, pageSizeKB)
-
-	for _, o := range optionSet.List() {
-		if strings.HasPrefix(o, "o=") {
-			v := o[2:]
-			if allowedOptions[v] {
-				csiMountOptions = append(csiMountOptions, v)
-			} else {
-				klog.Warningf("got invalid mount option %q. Will discard invalid options and continue to mount.", v)
-			}
-			optionSet.Delete(o)
-		}
-
-		if readAheadKB := readAheadKBMountFlagRegex.FindStringSubmatch(o); len(readAheadKB) == 2 {
-			// There is only one matching pattern in readAheadKBMountFlagRegex
-			// If found, it will be at index 1
-			readAheadKBInt, err := strconv.ParseInt(readAheadKB[1], 10, 0)
-			if err != nil {
-				return nil, nil, nil, 0, fmt.Errorf("invalid read_ahead_kb mount flag %q: %w", o, err)
-			}
-			if readAheadKBInt < 0 {
-				return nil, nil, nil, 0, fmt.Errorf("invalid negative value for read_ahead_kb mount flag: %q", o)
-			}
-			sysfsBDI[readAheadKBMountFlag] = readAheadKBInt
-			optionSet.Delete(o)
-		}
-
-		if maxReqSize := nodeFuseMaxRequestLimitKBMountFlagRegex.FindStringSubmatch(o); len(maxReqSize) == 2 {
-			nodeFuseMaxRequestLimitKB, err := strconv.ParseInt(maxReqSize[1], 10, 0)
-			if err != nil {
-				return nil, nil, nil, 0, fmt.Errorf("invalid %s mount flag %q: %w", nodeFuseMaxRequestLimitKBMountFlag, o, err)
-			}
-			if nodeFuseMaxRequestLimitKB < 0 {
-				return nil, nil, nil, 0, fmt.Errorf("invalid negative value for %s mount flag: %q", nodeFuseMaxRequestLimitKBMountFlag, o)
-			}
-			maxAllowedSizeKB := maxFuseMaxPagesLimit * pageSizeKB
-			if nodeFuseMaxRequestLimitKB > maxAllowedSizeKB {
-				return nil, nil, nil, 0, fmt.Errorf("invalid value for %s mount flag %q: exceeds maximum allowed limit of %d KiB on this host", nodeFuseMaxRequestLimitKBMountFlag, o, maxAllowedSizeKB)
-			}
-			// Convert the requested size in KB to pages. We use integer ceiling division
-			// to round up to the next page if the requested size is not a multiple of the
-			// page size. This ensures we can accommodate the full requested buffer size in
-			// a single fuse request.
-			fuseMaxPagesLimit = util.CeilDiv64(nodeFuseMaxRequestLimitKB, pageSizeKB)
-			optionSet.Delete(o)
-		}
-	}
-
-	return csiMountOptions, optionSet.List(), sysfsBDI, fuseMaxPagesLimit, nil
 }
 
 func checkForKernelReader(options []string) bool {
